% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data3CIA}
\alias{data3CIA}
\title{Individual Participant Data Meta-Analysis of COPD Patients}
\format{
A data frame the following columns:
\itemize{
\item \code{months}, length of follow-up in months;
\item \code{status}, indicator of the status at the end of follow-up (where
0: alive, 1: death);
\item \code{age}, age in years;
\item \code{fev1pp}, FEV1 measurement;
\item \code{mmrc}, dyspnea score (mMRC);
\item \code{cohort}, study indicator.
}
}
\usage{
data3CIA
}
\description{
A random subset of the 3CIA database of COPD patients.
}
\examples{
data("data3CIA", package = "stmepred")
}
\references{
\itemize{
\item J. B. Soriano, B. Lamprecht, A. S. Ramirez et al. Mortality prediction in
chronic obstructive pulmonary disease comparing the GOLD 2007 and 2011 staging
systems: A pooled analysis of individual patient data. The Lancet Respiratory
Medicine, 3(6):443--450, 2015. \url{doi:10.1016/S2213-2600(15)00157-5}
}
}
\keyword{datasets}
