% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data3Lsim}
\alias{data3Lsim}
\title{Simulated Three-Level Data}
\format{
A data frame the following columns:
\itemize{
\item \code{hospital_id}, ID of the 3^rd^ level of hierarchy (highest);
\item \code{provider_id}, ID of the 2^nd^ level of hierarchy;
\item \code{patient_id}, ID of the 1^st^ level of hierarchy (lowest);
\item \code{X1}, continuous covariate recorded at a patient-level;
\item \code{X2}, a second continuous covariate recorded at a patient-level;
\item \code{X3}, a binary covariate recorded at a patient-level;
\item \code{t}, time to event data;
\item \code{d}, event indicator variable, where 1 = event and 0 = censoring.
}
}
\usage{
data3Lsim
}
\description{
A simulated dataset with a three-levels hierarchical structure and
a time to event outcome.
}
\examples{
data("data3Lsim", package = "stdmest")
}
\keyword{datasets}
