% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_e.R
\name{read_e}
\alias{read_e}
\title{Read Estimation Results From Stata}
\usage{
read_e(path)
}
\arguments{
\item{path}{Path of .xlsx file containing \code{mestreg} estimation results as
exported by \code{mestreg_export} in Stata (bundled with the \code{stdmest}
command).}
}
\value{
A list with two elements, one for \code{e(b)} (named \code{eb}) and one for \code{e(V)}
(named \code{eV}).
}
\description{
Read Estimation Results From Stata
}
