% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stata.model.matrix.R
\name{Stata.model.matrix}
\alias{Stata.model.matrix}
\title{Create a Model Matrix Compatible With Stata}
\usage{
Stata.model.matrix(..., eb)
}
\arguments{
\item{...}{Everything that is passed to \code{\link[stats:model.matrix]{stats::model.matrix()}}, e.g., model
formula and data. Ideally, create indicator (one-hot encoding) variables
by hand, thus avoiding calling, e.g., \code{factor()}.}

\item{eb}{Fitted model coefficients, i.e., after using \code{\link[=read_e]{read_e()}}.}
}
\value{
A model matrix.
}
\description{
Create a Model Matrix Compatible With Stata
}
