% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdmest-package.R
\docType{package}
\name{stdmest-package}
\alias{stdmest-package}
\title{Regression Standardisation for Hierarchical Survival Models}
\description{
Description goes here...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RedDoorAnalytics/stdmest-r}
  \item Report bugs at \url{https://github.com/RedDoorAnalytics/stdmest-r/issues}
}

}
\author{
Alessandro Gasparini (alessandro.gasparini@reddooranalytics.se)
}
\keyword{internal}
