% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdmest.R
\name{stdmest}
\alias{stdmest}
\title{Regression Standardisation for Hierarchical Survival Models}
\usage{
stdmest(
  t,
  X,
  beta,
  Sigma,
  b,
  bse,
  bref = 0,
  brefse = 0,
  contrast = FALSE,
  distribution,
  conf.int = FALSE,
  B = 100,
  cimethod = "percentile",
  alpha = 0.05
)
}
\arguments{
\item{t}{A vector or single-column matrix with time points at which standardised
survival (or contrasts thereof) is to be computed.}

\item{X}{Fixed-effects design matrix, likely produced by \code{\link[=Stata.model.matrix]{Stata.model.matrix()}},
including all fixed effects to standardised over.}

\item{beta}{Estimated fixed-effects, likely produced by \code{\link[=read_e]{read_e()}}.}

\item{Sigma}{Estimated variance-covariance matrix, likely produced by \code{\link[=read_e]{read_e()}}.}

\item{b}{Value of the random intercept to fix. Can be a vector if multiple
hierarchical levels are present.}

\item{bse}{Standard error of \code{'b'}, required for the algorithm used when computing
confidence intervals. Can be a vector if multiple hierarchical levels are
present.}

\item{bref}{Value of the random intercept to fix as reference level. Defaults
to 0. Required when the argument \code{contrast = TRUE}. Can be a vector if multiple
hierarchical levels are present.}

\item{brefse}{Standard error of \code{'bref'}, required for the algorithm used when computing
confidence intervals. Can be a vector if multiple hierarchical levels are
present.}

\item{contrast}{Should a contrast be computed (e.g., survival difference comparing
\code{'b'} vs \verb{'bref})? Defaults to \code{TRUE}.}

\item{distribution}{Baseline hazard distribution. Possible values are \code{"exponential"}
and \code{"weibull"}.}

\item{conf.int}{Should confidence intervals for any standardised quantity be
computed? Defaults to \code{FALSE}, as the algorithm is computationally intensive.}

\item{B}{How many replications should the algorithm for computing confidence
intervals use? Defaults to 100.}

\item{cimethod}{Method used to obtain confidence intervals. Possible values are
the percentile method (\code{cimethod = "percentile"}, the default) or the normal
approximation method (\code{cimethod = "normal"}).}

\item{alpha}{Confidence level. Defaults to 0.05 for 95\% confidence intervals.}
}
\description{
This can be used to compute standardised survival probabilities
(over the case-mix covariates) while fixing the value of one (or more)
random intercepts.
}
