% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdmestm.R
\name{stdmestm}
\alias{stdmestm}
\title{Partially Marginal Regression Standardisation for Hierarchical Survival Models}
\usage{
stdmestm(
  t,
  X,
  beta,
  Sigma,
  b,
  bse,
  bref = 0,
  brefse = 0,
  varmargname,
  contrast = FALSE,
  distribution,
  conf.int = FALSE,
  B = 100,
  cimethod = "percentile",
  alpha = 0.05,
  nk = 7
)
}
\arguments{
\item{t}{A vector or single-column matrix with time points at which standardised
survival (or contrasts thereof) is to be computed.}

\item{X}{Fixed-effects design matrix, likely produced by \code{\link[=Stata.model.matrix]{Stata.model.matrix()}},
including all fixed effects to standardised over.}

\item{beta}{Estimated fixed-effects, likely produced by \code{\link[=read_e]{read_e()}}.}

\item{Sigma}{Estimated variance-covariance matrix, likely produced by \code{\link[=read_e]{read_e()}}.}

\item{b}{Value of the random intercept to fix. Must be a single scalar.}

\item{bse}{Standard error of \code{'b'}, required for the algorithm used when computing
confidence intervals. Must be a single scalar.}

\item{bref}{Value of the random intercept to fix as reference level. Defaults
to 0. Required when the argument \code{contrast = TRUE}. Must be a single scalar.}

\item{brefse}{Standard error of \code{'bref'}, required for the algorithm used when computing
confidence intervals. Must be a single scalar.}

\item{varmargname}{Name/label of the coefficient in 'beta' denoting the variance of
the random intercept to marginalise over (random intercept assumed assumed to
be centered on the value of zero).}

\item{contrast}{Should a contrast be computed (e.g., survival difference comparing
\code{'b'} vs \verb{'bref})? Defaults to \code{TRUE}.}

\item{distribution}{Baseline hazard distribution. Possible values are \code{"exponential"}
and \code{"weibull"}.}

\item{conf.int}{Should confidence intervals for any standardised quantity be
computed? Defaults to \code{FALSE}, as the algorithm is computationally intensive.}

\item{B}{How many replications should the algorithm for computing confidence
intervals use? Defaults to 100.}

\item{cimethod}{Method used to obtain confidence intervals. Possible values are
the percentile method (\code{cimethod = "percentile"}, the default) or the normal
approximation method (\code{cimethod = "normal"}).}

\item{alpha}{Confidence level. Defaults to 0.05 for 95\% confidence intervals.}

\item{nk}{Number of nodes used for the numerical integration procedure: larger
values yield more accurate results, at the cost of additional computational
complexity. Defaults to 7, which (anecdotally) yielded a good trade-off
between accuracy and computational cost.}
}
\description{
This can be used to compute standardised survival probabilities
(over the case-mix covariates) while fixing the value of one of the random
intercepts while marginalising over the others.
Note that currently only three-levels models are supported, e.g., with a
fixed random intercept and a random intercept that we marginalise over.
For comparison, \code{\link[=stdmest]{stdmest()}} fixes all random intercepts instead.
}
