% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mestreg.R
\name{read_e}
\alias{read_e}
\title{Read Estimation Results From Stata}
\usage{
read_e(path_eb, path_eV)
}
\arguments{
\item{path_eb}{Path of .xlsx file containing \code{e(b)} estimation results as
exported by \code{putexcel} in Stata.}

\item{path_eV}{Path of .xlsx file containing \code{e(V)} estimation results as
exported by \code{putexcel} in Stata.}
}
\value{
A list with two elements, one for \code{e(b)} (named \code{eb}) and one for \code{e(V)}
(named \code{eV}).
}
\description{
Read Estimation Results From Stata
}
