% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_e.R
\name{read_e}
\alias{read_e}
\title{Read Estimation Results From Stata}
\usage{
read_e(path)
}
\arguments{
\item{path}{Path of .xlsx file containing \code{mestreg} or \code{stmixed} estimation
results as exported by \code{modexpt} in Stata (bundled with the \code{stdmest}
command).}
}
\value{
A list with different components (depending on whether we are reading
in results from \code{mestreg} or \code{stmixed}. Common elements will be \code{eb} (for
\code{e(b)}) and \code{eV} (for \code{e(V)}), for the estimated coefficients and their
variance-covariance matrix, among others. Check the input .xlsx file and
\verb{ereturn list} in Stata (after \code{mestreg}, \code{stmixed}) for more details.
}
\description{
Read Estimation Results From Stata
}
